fix: fix config and body parser bugs
