+++ src/cluster.ts
@@ -1,6 +1,6 @@
 import cluster from "cluster";
 import os from 'os';
-import { createServer } from "./index.js";
+import { createServer } from "./worker.js";
 import dotenv from 'dotenv';
 dotenv.config();
 const PORT: number = Number(process.env.PORT) || 3000;
